refactor: move core files to app folder
