+++ src/main.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import vuetify from "@/plugins/vuetify";
+import App from "@/app/app.vue";
+import router from "@/app/router";
+import vuetify from "@/app/plugins/vuetify";
 import "material-design-icons-iconfont/dist/material-design-icons.css";
 
 Vue.config.productionTip = false;
